Moved TextWidgetUtils to widgets/utilities
